+++ src/options.tsx
@@ -2,49 +2,49 @@ import React, { useEffect, useState } from "react";
 import ReactDOM from "react-dom/client";
 import "semantic-ui-css/semantic.min.css";
 import { Popup } from "semantic-ui-react";
-import { getShareURLPageScript } from "./worker.ts";
+import { getSaveURLPageScript } from "./worker.ts";
 
-const ShareURLPresets = [
+const SaveURLPresets = [
   ["Raindrop", "https://app.raindrop.io/add?link={url}&title={title}"],
   ["Pinboard", "https://pinboard.in/add?url={url}&title={title}"],
   ["Pocket", "https://getpocket.com/edit?url={url}"],
 ];
 
 const PageInfoOptions = () => {
-  const [shareIcon, setShareIcon] = useState<string | null>(null);
-  const [shareURLTemplate, setShareURLTemplate] = useState<string | null>(null);
-  const [shareURLInBackground, setShareURLInBackground] = useState<
+  const [saveURLIcon, setSaveURLIcon] = useState<string | null>(null);
+  const [saveURLTemplate, setSaveURLTemplate] = useState<string | null>(null);
+  const [saveURLInBackground, setSaveURLInBackground] = useState<
     boolean | null
   >(null);
-  const updateShareURLTemplate = (newShareURLTemplate: string | null) => {
-    setShareURLTemplate(newShareURLTemplate);
-    if (newShareURLTemplate !== null) {
-      shareIcon ?? setShareIcon("bookmark");
-      if (getShareURLPageScript(newShareURLTemplate) !== null) {
-        shareURLInBackground ?? setShareURLInBackground(false);
+  const updateSaveURLTemplate = (newSaveURLTemplate: string | null) => {
+    setSaveURLTemplate(newSaveURLTemplate);
+    if (newSaveURLTemplate !== null) {
+      saveURLIcon ?? setSaveURLIcon("bookmark");
+      if (getSaveURLPageScript(newSaveURLTemplate) !== null) {
+        saveURLInBackground ?? setSaveURLInBackground(false);
       } else {
-        setShareURLInBackground(null);
+        setSaveURLInBackground(null);
       }
     }
   };
 
-  const handleChangeShareIcon = (
+  const handleChangeSaveURLIcon = (
     event: React.ChangeEvent<HTMLInputElement>,
   ) => {
-    setShareIcon(event.target.value || null);
+    setSaveURLIcon(event.target.value || null);
   };
-  const handleChangeShareURLTemplate = (
+  const handleChangeSaveURLTemplate = (
     event: React.ChangeEvent<HTMLInputElement>,
-  ) => updateShareURLTemplate(event.target.value || null);
-  const handleChangeShareURLInBackground = (
+  ) => updateSaveURLTemplate(event.target.value || null);
+  const handleChangeSaveURLInBackground = (
     event: React.ChangeEvent<HTMLInputElement>,
-  ) => setShareURLInBackground(event.target.checked);
+  ) => setSaveURLInBackground(event.target.checked);
   const handleSave = () => {
     chrome.storage.sync.set(
       {
-        shareIcon,
-        shareURLTemplate,
-        shareURLInBackground,
+        saveURLIcon,
+        saveURLTemplate,
+        saveURLInBackground,
       },
       () => window.close(),
     );
@@ -60,14 +60,14 @@ const PageInfoOptions = () => {
 
     chrome.storage.sync.get(
       {
-        shareIcon: null,
-        shareURLTemplate: null,
-        shareURLInBackground: false,
+        saveURLIcon: null,
+        saveURLTemplate: null,
+        saveURLInBackground: false,
       },
-      ({ shareIcon, shareURLTemplate, shareURLInBackground }) => {
-        setShareIcon(shareIcon);
-        setShareURLTemplate(shareURLTemplate);
-        setShareURLInBackground(shareURLInBackground);
+      ({ saveURLIcon, saveURLTemplate, saveURLInBackground }) => {
+        setSaveURLIcon(saveURLIcon);
+        setSaveURLTemplate(saveURLTemplate);
+        setSaveURLInBackground(saveURLInBackground);
       },
     );
   }, []);
@@ -86,7 +86,7 @@ const PageInfoOptions = () => {
         >
           <fieldset>
             <legend className="block mb-2 text-base font-medium">
-              Share Icon
+              Save URL Icon
             </legend>
 
             <ul className="flex">
@@ -95,18 +95,18 @@ const PageInfoOptions = () => {
                   <li className="flex items-center mr-4">
                     <input
                       type="radio"
-                      id="shareIcon"
-                      name="shareIcon"
+                      id="saveURLIcon"
+                      name="saveURLIcon"
                       value={value || ""}
-                      checked={shareIcon == value}
-                      onChange={handleChangeShareIcon}
+                      checked={saveURLIcon == value}
+                      onChange={handleChangeSaveURLIcon}
                       className="w-4 h-4 border-gray-300 focus:ring-2 focus:ring-blue-300 dark:focus:ring-blue-600 dark:focus:bg-blue-600 dark:bg-gray-700 dark:border-gray-600"
                     />
                     <label
                       className={`ml-1 text-sm font-medium text-gray-900 dark:text-gray-300`}
                     >
                       {value ? (
-                        <i className={`share-icon ${value} icon`} />
+                        <i className={`save-url-icon ${value} icon`} />
                       ) : (
                         "Disabled"
                       )}
@@ -119,10 +119,10 @@ const PageInfoOptions = () => {
 
           <fieldset>
             <label
-              htmlFor="shareURLTemplate"
+              htmlFor="saveURLTemplate"
               className="block mb-2 text-base font-medium"
             >
-              Share URL Template
+              Save URL Service
             </label>
 
             <select
@@ -130,21 +130,21 @@ const PageInfoOptions = () => {
               onChange={(event: React.ChangeEvent<HTMLSelectElement>) => {
                 const value = event.target.value;
                 if (value) {
-                  if (shareIcon === null) setShareIcon("bookmark");
-                  setShareURLTemplate(value);
+                  if (saveURLIcon === null) setSaveURLIcon("bookmark");
+                  setSaveURLTemplate(value);
                 }
               }}
             >
               <option
                 value=""
                 selected={
-                  !ShareURLPresets.some(([_, url]) => shareURLTemplate === url)
+                  !SaveURLPresets.some(([_, url]) => saveURLTemplate === url)
                 }
               >
                 Custom
               </option>
-              {ShareURLPresets.map(([label, url]) => (
-                <option value={url} selected={shareURLTemplate === url}>
+              {SaveURLPresets.map(([label, url]) => (
+                <option value={url} selected={saveURLTemplate === url}>
                   {label}
                 </option>
               ))}
@@ -155,11 +155,11 @@ const PageInfoOptions = () => {
                 trigger={
                   <input
                     type="text"
-                    id="shareURLTemplate"
-                    name="shareURLTemplate"
-                    value={shareURLTemplate || undefined}
+                    id="saveURLTemplate"
+                    name="saveURLTemplate"
+                    value={saveURLTemplate || undefined}
                     placeholder="https://.../add?url={url}&title={title}"
-                    onChange={handleChangeShareURLTemplate}
+                    onChange={handleChangeSaveURLTemplate}
                     className="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2.5"
                   />
                 }
@@ -169,21 +169,21 @@ const PageInfoOptions = () => {
               />
             </div>
 
-            {shareURLInBackground !== null && (
+            {saveURLInBackground !== null && (
               <div className="flex items-center mt-2">
                 <input
                   type="checkbox"
-                  id="shareURLInBackground"
-                  name="shareURLInBackground"
-                  checked={shareURLInBackground}
-                  onChange={handleChangeShareURLInBackground}
+                  id="saveURLInBackground"
+                  name="saveURLInBackground"
+                  checked={saveURLInBackground}
+                  onChange={handleChangeSaveURLInBackground}
                   className="w-4 h-4 text-blue-600 bg-gray-100 border-gray-300 rounded focus:ring-blue-500 dark:focus:ring-blue-600 dark:ring-offset-gray-800 dark:focus:ring-offset-gray-800 focus:ring-2 dark:bg-gray-700 dark:border-gray-600"
                 />
                 <label
-                  htmlFor="shareURLInBackground"
+                  htmlFor="saveURLInBackground"
                   className="ml-1 text-sm font-medium text-gray-900 dark:text-gray-300"
                 >
-                  Open the share URL popup in background and auto-close it.
+                  Open the save URL popup in background and auto-close it.
                 </label>
               </div>
             )}